Fix inverse stereo projection on an ellipsoid.
The problem is that the setup function doesn't initialise all of the
values that it needs for the ellipsoid, inverse, equatorial case.

In the forward and spherical inverse cases, the equatorial case
(|phi0| < 1e-10) is a simplification of the oblique case, and requires
fewer parameters. However, for the ellipsoid inverse case (e_inverse),
the oblique calculation is used, but the setup function doesn't
initialise all of the parameters.