Fix extraneous parentheses warning.
This warning indicates maybe we wanted to use assignment and added the
extra parentheses to silence that *other* warning about assignment in a
conditional. However, we really want a conditional, and the parentheses
are due to it being in a macro, so using a yoda conditional avoids the
ambiguity and silences the warning.