Fix the constants used for degree/radian conversions.
In dmstor.c an inaccurate constant was used for pi/180.  This meant that
90deg converted to radians and back to degrees gave a result which was
larger than 90deg.  The constant has been replaced by DEG_TO_RAD
(defined in proj_api.h).

In rtodms.c, the trailing digits in CONV were wrong; the constant has
been truncated so that all the digits are corrent.