Take 2 of the fix to dmstor.
I'm making this pull request soon after the release of 4.9.2.  It will
cause the results that people get out of proj.4 to change very slightly.
If there are problems, we'll get a chance to iron them out well before
the next release.

The important change is to use DEG_TO_RAD for degree to radian
conversions in dmstor.c instead of the slightly inaccurate number used
earlier.

This necessitates a change to geod_interface.c (which previously had to
work aroung the previous bad behavior).  PJ_aeqd.c now does conversions
in a compatible manner.

In src/proj_api.h, DEG_TO_RAD and RAD_TO_DEG are both given to 17
significant figures (this is just a cosmetic change).

I've "fixed" the testvarious tests so that they still pass (on my system
at least).  Everyone should be suitably skeptical of my fixes.

(1) Minor changes to "Test bug 244" and only ask for nanometer (instead
of picometer) accuracy on positions.

(2) 2 lon_wrap tests now return 0dE instead of 360dE (now it's tight?)

(3) The results for the forward healpix projection of (-180, +90) and
(-180, -90) are now different.  I have put the new values into
tv_out.dist.  I'm fairly confident that the new values are OK, since
this projection has various cuts which meet at the poles.  It would be
good if someone who knows about this projection can verify this.