Increase maximum path size for grid shift loading
On systems with long paths (e.g. iOS, which uses UUIDs for bundle
directories), the existing maximum path length of 128 resulted in path
truncation, and caused the grid shift file loading to fail.

This change increases the buffer size for the path to 1024 chars,
enabling the nadgrids file to be loaded on such systems.