Fix +wgs84 parameters for srid 3844 (Romania)
The Parameters are taken from EPSG 8.9

Relation with https://trac.osgeo.org/postgis/ticket/1851
and https://trac.osgeo.org/geotiff/ticket/52

Quoting from the Postgis ticket:

"The transformation parameters between S42 and WGS84 datums wihch are listed inside the EPSG:3844 definition
(TOWGS84[33.4,-146.6,-76.3,-0.359,-0.053,0.844,-0.84]) are valid for Poland, not for Romania.
You can see this in the EPSG database: Coordinate Transformation with code 1645: Area of use = Poland - onshore.
The correct parameters for datum transformation in the case of EPSG:3844 are listed in the EPSG database with coordinate transformation code = 15995."