selftest function stubs
Added stubs for selftest functions for all projections supported.

For now, the stub collection has been addad at the bottom of PJ_aea.c

Each time a real selftest function is added to a projection source file,
the stub should be removed.

Stubbing the stuff makes sure a selftest function exists for all
projections, and hence allows us to build a list of all selftest
functions using the X-macro trick (with PROJ_HEAD as X-macro) used for
building pj_list[] in pj_init.c.

Actually, the plan is to add a selftest entry directly in the struct
PJ_LIST, and let pj_init build the list directly as a part of pj_list[]