Improved pipeline output for proj
The pipeline can now signal to the proj executable, that it provides
angular output, and hence ask proj to avoid treating output as linear
units (and hence multiply by the semimajor axis length).

Proj.c now listens to what the pipeline manager says.

This needs more work to support the -I (invproj) option: We need to
specify both expected input and provided output, and to provide an
override facility for the user.