Geodetic-to-geocentric revisited
As part of the pipeline work, reimplement the geodetic-to-geodetic
functionality in proper 3D PJ style.

The existing PJ_geodetic is mostly a stub, deferring the actual work to
behind-the-scenes code in pj_transform.c.

For datum shift pipelines, we want a proper 3D, accessible version. This
is implemented in PJ_cart.c