COORDINATE renamed to TRIPLET; Added minimalist API defined in "proj.h"
Apart form being the generic header for the TRIPLET union, which is
useful for geodetic applications in general, proj.h also provides access
to the minimalistic API consisting of:
pj_init_plus
pj_apply
pj_free
i.e. the bare essentials needed for using proj.4 in a TRIPLET based
workflow.

I plan to add a few things to this, while keeping it both simple and
useful