Added missing DMS triplet + minor corrections in response to a review by Micah Cochran
Silence MSVC warnings about promotion of int to double, by changing 180
to Fortran style 180.0.

As the promotion is exactly as specified in the standard it is slightly
odd to warn about it.

But minimizing the number of warnings makes it easier to focus on
actually helpful warnings about incidentally introduced potential
dangers.