Added vertical gridshift "projection".
Until now gridshifts has not been working with the new API in
proj.h since parsing of +nadgrids and +geoidgrids is build
into pj_transform(). This commit introduces the possibility to
do vertical gridshift with the pipeline API. The vgridshift
kernel is a simple wrapper for pj_apply_vgridshift() that is
also used by pj_transform().