Added horizontal gridshift driver.
Until now gridshifts has not been working with the new API in
proj.h since parsing of +nadgrids and +geoidgrids is build
into pj_transform(). This commit introduces the possibility to
do horizontal gridshift with the pipeline API. The hgridshift
driver is a simple wrapper for the funtions in pj_apply_gridshift.c
that is also used by pj_transform().