Horisontal and vertical gridshift drivers
Until now gridshifts has not been working with the new API in
proj.h since parsing of +nadgrids and +geoidgrids is build
into pj_transform(). This commit introduces the possibility to
do both horizontal and vertical gridshift with the pipeline API. 
The vgridshift and hgridshift kernels are simple wrappers for 
pj_apply_gridshift3() and pj_apply_vgridshift() that are
also used by pj_transform().

Introduced in PR #492.