Removed unwanted call to pj_log_level().
When running the self-test with logging was turned off globally when
the PJ_cart test was run. As a consequence all test functions
called after PJ_cart was not able to output debugging information.
In most cases this would go by unnoticed but when calling proj with
PROJ_DEBUG=1 (or higher) all debugging output whould get suppressed.

This commit removes the call to pj_log_level that turns of debugging
output in the self-test.