Extended Helmert transformation to 14-parameters
Extended Helmert transformation to 14-parameters.

This commit extends the Helmert transformationto the fourth dimension and
enables spatio-temporal datum shifts in PROJ.4. On top of the usual 7
parameters (+x, +y, +z, +s, +rx, +ry, +rz) the rates of change of the seven
parameters can now be used as well. The new parameters are called +dx,
+dy, +dx, +ds, +drx, +dry and +drz. To keep track of the datum epoch and
coordinate epoch two additional parameters have been added to the
Helmert transformation, one of which is mandatory in the 14-parameter
case. The mandatory datum epoch is controlled with +epoch (given in
YYYY.yyyy format) and the coordinate, or observation, epoch is either
controlled in the proj-string with +tobs or by using 4D-coordinates
when transformating coordinates with pj_trans().

See the test functions for examples of how to set up the transformation
with 14 parameters.