4-parameter Helmert.
Completing the Helmert driver with the 4-parameter shift that handles
the 2D transformation. The implementation is written in such a way that
not only 2D-points but also 3D- and 4D-points can be transformed with
the 4-parameter Helmert. The four parameters that can be set in this
mode are +x, +y (translations), +s (scale) and +theta (rotation). The
presence of the +theta parameter activates the 2D-helmert code,
irregardless of the input data's dimensions.

The units are meters for the translations as in all the other versions
of the Helmert transform. The rotation unit is arcseconds. The units of
the scale differ from the 3-, 7- and 14-parameter shift where the unit
is ppm. Here it is instead given directly and is as such unitless.

The 4-parameter case can also be extended to an 8-parameter shift in
the same way as the 7-parameter shift extens to the 14-parameter shift.
This might be a bit silly and will probably never be used, but
nonetheless, I have included it for the sake of completeness. The rates
of change are givens as +dx, +dy, +ds and +dtheta.