Expose pj_factors and pj_deriv in public API. (#546)
Expose pj_factors and pj_deriv in public API.

proj_derivatives() and proj_factors() has been
introduced to the public proj.h API. The new functions are simply
wrappers for the original functions. The h argument of the two functions
has been omitted in the new wrappers as it does not seem to be used very
much (if at all?).