Add tests for C99 math functions, hypot, atanh, cbrt, etc.
This tests whether the functions are declared in <math.h>.  If they are,
then -DHAVE_C99_MATH=1 is added to the C flags.  The intention is that
this flag is only seen when building proj.4 and shouldn't be referenced
in any of the installed include files.  The next update to geodesic.c
will use this flag.

Left unaddressed is what to do if HAVE_C99_MATH is 0.  The strategy in
geodesic.c is to assume that the missing functions will need to be
defined explicitly.  A less safe alternative is to assume that the
functions are in fact available in libm and that all that needs to be
done is to declare the functions.