Allow nested pipelines.
The previous behaviour was to quit pipeline initialization when
encountering a nested pipeline definition. With this commit that
behaviour is changed so that nested pipelines are expanded into one big
pipeline. Example:

    +proj=pipeline +step +proj=pipeline +step +proj=merc

is turned into

    +proj=pipeline +step +proj=merc

This is useful in init-files where steps in complicated transformations
can be grouped in "sub-pipelines". These "sub-pipelines" can then be
used as individual steps in a larger and more complicated pipeline.