Ensure C89 compliance and enable more warnings (#650)
* Ensure C89 compliance and enable more warnings

According to the contributing guidelines, the library is developed
strictly in ANSI C 89. However, this is not enforced.

Additionally, we enable more warnings: apart from -Wall and -Wextra we
enable a warning that makes sure all enumeration values are covered in
a switch statement.

When compiling with Clang, we also turn on the warnings -Wc99-extensions
and -Wc11-extensions.

* Enable all warnings that are used on Travis

This makes it easier to notice problems before pushing.