Add plate motion model parameters to ITRF init-files.
Parameters for the plate motion models (PMM) for ITRF2008 and ITRF2014
are added to the ITRF2008/2014 init-files. The PMMs allow coordinates
to be moved back and forward in time in plate fixed reference frames
such as GR96 in Greenland which is defined as ITRF94@1996.623.
Transforming an ITRF2014-coordinate to GR96 is done with:

+proj=pipeline +step +init=ITRF2014:NOAM   +t_epoch=1996.623 +t_obs=2017.584
               +step +init=ITRF2014:ITRF94 +t_obs=1996.623

where the first step transforms the coordinate back in time to
ITRF2014@1996.632 by using the ITRF2014 PMM. The second step transforms
the ITRF2014 coordinate to ITRF94.