Improve tests and add test files to automake setup.
Make sure that the gie files in test/gie and test/gigs are included
in the distribution tarball. Tests with gie are set up to run all
available files matching *.gie in the gie and gigs directories. For
this reason tests that are known to fail in gigs/ have been renamed
to *.gie_failing, so that "make check" can be run in-tree in local
copies of the git repository.