gie expect: ignore unspecd dims, improve reporting/builtins
In pipelines including a Helmert shift, we need to run the test
through the 4D transformation interface, even though the input
coordinate system may be 2D.

This can be enforced by appending "0 0" to the 2D coordinate pair
in the accept instruction, which is sufficiently recognizable to
be considered an idiom for selecting 4D.

On return, however, (i.e. in the expect instruction), the last
dimensions will contain garbage, and this garbage will be
compared with "0 0" when computing the deviation.

This obviously leads to nonsensical results, which this commit
repairs by zeroing all dimensions *not given* in expect, before
computing the deviation.

Additionally, the test tolerance for geo/cartesian roundtrip
precision has been relaxed from picometer to nanometer level.

These tests have shown to intermittently bomb, and as the pm
level tolerance is probably a leftover from when deviation was
computed in degrees, not meter (and hence a factor of 111000
more tight than intended at its introduction) relaxing it
by a factor of 1000 makes ample sense.

Also, two new features, introduced while debugging this case
has been left in the code:

- improved reporting, for verbosity levels higher than 2

- a "skip" instruction, forcing all remaining work to be
  skipped (i.e. run until something strange happens - then stop
  to handle debugging, while avoiding additional garbage)