Add tests to improve the code coverage.
Now all of geodesic.c is covered except for 3 lines (and 2 of these are
intentionally dead code).
    
This corresponds to tag v1.49.2-c in the GeographicLib code base.

While testing polygons which encircle the globe multiple times, I
uncovered a problem where the range of the area was not reduced to the
allowed range (either [0, area0) or (-area0/2, area0/2]) correctly.
Since the documentation explicity restricted the calculation of polygon
areas to simple polygons, we'll defer fixing this for now.  (However the
intention was always to handle the area "algebraically" so that, for
example, a "bowtie" has zero area.  So I will plan on fixing this for
1.50.)

Update copyright dates + NEWS.