Add section on PROJ resource files to docs [skip ci]
Added new section on PROJ resource files to the top level of the
documentation. Moved information on init-file from the 'usage'
chapter to the new section. Also moved pages on grids and HTPD to the
new resource files section. Finally links and information on the new
regional datumgrids packages is added.