Reformat doc pages for cct, cs2cs, geod and proj
The doc pages are reformatted to better use the features of the Sphinx
documentation system. All four pages has been formatted such that they
can serve as the basis for both the website and UNIX man-pages. This
will make it easier to keep the man-pages aligned with the main
documentation in the future.