Setup automatic man page generation
Create man pages from the Sphinx documentation. Ideally the man pages
are updated completely automatic when running make dist and make install
but that is a quite complicated setup. For now a script that builds the
man pages and commits them has been created. This is to be run as part
of the release process so that man pages are aligned with the most
recent version of the docs.