Don't set -std=c89 in CMakeLists.txt
This allows us to take advantage of newer features when they are
available. However, builds on Travis still use -std=c89 to ensure C89
compatibility. Locally this can be achieved with either

* cmake -DCMAKE_C_STANDARD=90 ..
* cmake -DCMAKE_C_FLAGS='-std=c89' ..
* C_FLAGS='-std=c89' cmake ..

We also reorder the warning flags: they are all part of the standard
build now, but -Werror is only applied on Travis.

Fixes #892.