Use Sphinx option directives to document PROJ options
Instead of listing the various projection options in a table we write
them as :option: tags instead. This makes it easier to write a more
elaborate description of the parameter. Additionally we can create
"master"-description of commonly used parameters such as +lon_0 and +k_0
and include them where necessary. This is handled in separate files, one
for each common parameters, that are stored in
docs/source/operations/options/.