Improve PDF generation of docs
Removed content sections from a bunch of pages since they clutter up the
PDF output and the same info is available in the sidebar of the webpage.

A few sections has been turned of for the PDF output, most notable the
front page of the webpage. It doesn't really fit in the PDF.