Handle double to int typecasts in ISEA better
Originally the code was doing double to int conversions like

    y = (int)(x + 0.5)

which results in rounding when typecasting. In an earlier attempt to
avoid buffer overflows in integer typecasts this was changed to

    y = lround(x + 0.5)

which doesn't give the origial results. We fix that here by instead
doing

    y = lround(x)

It is safe to so as long as x is positive.