Correct bad use of proj_dmstor in gie.
The previous commits were added way to prematurely. The code was not
properly tested and it turned out to do more bad than good. This commit
hopefully fixes that. And this time it is backed up by tests!

DMS style coordinates should now be fully functional in gie. Finger
crossed.