Renamed Web Mercator
More precise name of Web Mercator or Pseudo Mercator is WGS 84 Web Mercator / WGS 84 Pseudo Mercator. This is its name in EPSG:3857 because it is tied to ellipsoid WGS 84. 

Of course, equations of that map projection can be applied to any ellipsoid, but then it is not any more Web Mercator projection as was designed.