Remove dead code in fwd_finalize() and inv_prepare()
Those functions contain code specific of input != angular and output = angular
which to the best of my knowledge never happens in PROJ. Looking at that
code, I feel there are a number of errors in it, and anyway removing it shows
absolutely no change in the test suite, which shows it is unused. I've also
added exit(1) to verify that those code paths are never taken.

This was found while investigating the fix for
8cfc81380617ff4a17a06a97635f77c5e9ed7d5b